requirements.txt
python-telegram-bot==21.1.1
numpy==1.26.4
pandas==2.2.2
matplotlib==3.8.4
scikit-learn==1.5.0
requests==2.32.3
tqdm==4.66.4
torch==2.5.1 --index-url https://download.pytorch.org/whl/cu121